# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Prompt-to-Form API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthCheckResponse {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: string;
  email: string;
}

export type FormStatus = (typeof FormStatus)[keyof typeof FormStatus];

export const FormStatus = {
  draft: "draft",
  published: "published",
  closed: "closed",
} as const;

export interface Form {
  id: string;
  userId: string;
  title: string;
  description?: string | null;
  featureImageUrl?: string | null;
  slug: string;
  originalLanguage: string;
  preferredLanguage?: string | null;
  status: FormStatus;
  supportedLanguages: string[];
  responseLimit?: number | null;
  closesAt?: string | null;
  createdAt: string;
}

export type FormFieldFieldType =
  (typeof FormFieldFieldType)[keyof typeof FormFieldFieldType];

export const FormFieldFieldType = {
  short_text: "short_text",
  long_text: "long_text",
  single_choice: "single_choice",
  multi_choice: "multi_choice",
  rating: "rating",
  date: "date",
  email: "email",
  phone: "phone",
} as const;

export interface FormField {
  id: string;
  formId: string;
  orderIndex: number;
  fieldType: FormFieldFieldType;
  label: string;
  placeholder?: string | null;
  isRequired: boolean;
  optionsJson?: string[] | null;
}

export type FormTranslationTranslationsJson = { [key: string]: unknown };

export interface FormTranslation {
  id: string;
  formId: string;
  language: string;
  translationsJson: FormTranslationTranslationsJson;
  generatedAt: string;
}

export type SubmissionRawResponsesJson = { [key: string]: unknown };

export type SubmissionTranslatedResponsesJson = {
  [key: string]: unknown;
} | null;

export type SubmissionSentimentAnalysisJson = { [key: string]: unknown } | null;

export type SubmissionTranslationStatus =
  (typeof SubmissionTranslationStatus)[keyof typeof SubmissionTranslationStatus];

export const SubmissionTranslationStatus = {
  pending: "pending",
  done: "done",
  skipped: "skipped",
} as const;

export interface Submission {
  id: string;
  formId: string;
  respondentLanguage: string;
  rawResponsesJson: SubmissionRawResponsesJson;
  translatedResponsesJson?: SubmissionTranslatedResponsesJson;
  sentimentAnalysisJson?: SubmissionSentimentAnalysisJson;
  translationStatus: SubmissionTranslationStatus;
  submittedAt: string;
}

export type FormWithFieldsStatus =
  (typeof FormWithFieldsStatus)[keyof typeof FormWithFieldsStatus];

export const FormWithFieldsStatus = {
  draft: "draft",
  published: "published",
  closed: "closed",
} as const;

export interface FormWithFields {
  id: string;
  userId: string;
  title: string;
  description?: string | null;
  featureImageUrl?: string | null;
  slug: string;
  originalLanguage: string;
  preferredLanguage?: string | null;
  status: FormWithFieldsStatus;
  supportedLanguages: string[];
  responseLimit?: number | null;
  closesAt?: string | null;
  createdAt: string;
  fields: FormField[];
}

export interface PublicForm {
  id: string;
  title: string;
  description?: string | null;
  featureImageUrl?: string | null;
  slug: string;
  originalLanguage: string;
  supportedLanguages: string[];
  fields: FormField[];
  translations: FormTranslation[];
}

export type FormStatsLanguageBreakdown = { [key: string]: number };

export interface FormStats {
  totalResponses: number;
  languageBreakdown: FormStatsLanguageBreakdown;
  recentResponses: number;
}

export type DashboardSummaryWeeklyTrendItem = {
  date: string;
  count: number;
};

export type DashboardSummaryTopLanguagesItem = {
  language: string;
  count: number;
};

export type DashboardSummaryMostActiveForm = {
  id: string;
  title: string;
  responseCount: number;
} | null;

export interface DashboardSummary {
  totalForms: number;
  publishedForms: number;
  draftForms: number;
  totalResponses: number;
  recentForms: Form[];
  weeklyTrend: DashboardSummaryWeeklyTrendItem[];
  topLanguages: DashboardSummaryTopLanguagesItem[];
  mostActiveForm?: DashboardSummaryMostActiveForm;
}

export interface UpdateMeBody {
  /** @minLength 6 */
  password?: string;
  preferredLanguage?: string | null;
}

export type CreateFieldBodyFieldType =
  (typeof CreateFieldBodyFieldType)[keyof typeof CreateFieldBodyFieldType];

export const CreateFieldBodyFieldType = {
  short_text: "short_text",
  long_text: "long_text",
  single_choice: "single_choice",
  multi_choice: "multi_choice",
  rating: "rating",
  date: "date",
  email: "email",
  phone: "phone",
} as const;

export interface CreateFieldBody {
  fieldType: CreateFieldBodyFieldType;
  label: string;
  placeholder?: string | null;
  isRequired: boolean;
  optionsJson?: string[] | null;
  orderIndex?: number;
}

export interface CreateFormBody {
  title: string;
  description?: string | null;
  featureImageUrl?: string | null;
  originalLanguage: string;
  fields: CreateFieldBody[];
}

export type UpdateFormBodyStatus =
  (typeof UpdateFormBodyStatus)[keyof typeof UpdateFormBodyStatus];

export const UpdateFormBodyStatus = {
  draft: "draft",
  published: "published",
  closed: "closed",
} as const;

export interface UpdateFormBody {
  title?: string;
  description?: string | null;
  featureImageUrl?: string | null;
  status?: UpdateFormBodyStatus;
  preferredLanguage?: string | null;
  supportedLanguages?: string[];
  responseLimit?: number | null;
  closesAt?: string | null;
}

export interface PublishFormBody {
  languages: string[];
}

export interface PublishFormResult {
  success: boolean;
  failedLanguages: number;
  form: Form;
}

export interface UpdateFieldBody {
  label?: string;
  placeholder?: string | null;
  isRequired?: boolean;
  optionsJson?: string[] | null;
  orderIndex?: number;
}

export interface ReorderFieldsBody {
  fieldIds: string[];
}

export type SubmitFormBodyResponses = { [key: string]: unknown };

export interface SubmitFormBody {
  respondentLanguage: string;
  responses: SubmitFormBodyResponses;
}

export interface GenerateFormBody {
  prompt: string;
  /** BCP-47 language code to generate the form in (e.g. "en", "fr"). Defaults to auto-detect from prompt. */
  language?: string;
}

export type GenerateFormResultForm = {
  title: string;
  description?: string | null;
  featureImageUrl?: string | null;
  fields: CreateFieldBody[];
};

export interface GenerateFormResult {
  form: GenerateFormResultForm;
  detectedLanguage: string;
}

export interface FormTemplate {
  id: string;
  formId: string;
  userId: string;
  title: string;
  description?: string | null;
  featureImageUrl?: string | null;
  category: string;
  isPublic: boolean;
  useCount: number;
  createdAt: string;
  fieldCount: number;
  fields?: FormField[];
}

export interface SaveTemplateBody {
  title: string;
  description?: string | null;
  featureImageUrl?: string | null;
  category?: string;
  isPublic: boolean;
}

export type GenerateFormSentimentSummary200 = {
  summary?: string;
};

export type ReorderFields200 = {
  success?: boolean;
};

export type ListTemplatesParams = {
  /**
   * Optional category filter
   */
  category?: string;
};

export type SubmitForm201 = {
  id?: string;
};