# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Prompt-to-Form API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /forms/{id}/template:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getFormTemplate
      tags: [templates]
      summary: Get template info for a specific form
      responses:
        "200":
          description: Template entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormTemplate"
        "404":
          description: Not a template

    post:
      operationId: saveFormTemplate
      tags: [templates]
      summary: Save form as a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveTemplateBody"
      responses:
        "200":
          description: Template saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormTemplate"
        "404":
          description: Form not found

    delete:
      operationId: removeFormTemplate
      tags: [templates]
      summary: Remove template status from a form
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List available templates
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Optional category filter
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FormTemplate"
...
  /templates/{id}/use:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: useTemplate
      tags: [templates]
      summary: Create a new form from a template
      responses:
        "201":
          description: New form created from template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormWithFields"
        "404":
          description: Template not found
...
components:
  schemas:
...
    FormField:
      type: object
      properties:
        id:
          type: string
        formId:
          type: string
        orderIndex:
          type: integer
        fieldType:
          type: string
          enum: [short_text, long_text, single_choice, multi_choice, rating, date, email, phone]
        label:
          type: string
        placeholder:
          type: string
          nullable: true
        isRequired:
          type: boolean
        optionsJson:
          type: array
          items:
            type: string
          nullable: true
      required: [id, formId, orderIndex, fieldType, label, isRequired]
...
    FormWithFields:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        featureImageUrl:
          type: string
          nullable: true
        slug:
          type: string
        originalLanguage:
          type: string
        preferredLanguage:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, closed]
        supportedLanguages:
          type: array
          items:
            type: string
        responseLimit:
          type: integer
          nullable: true
        closesAt:
          type: string
          nullable: true
        createdAt:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FormField"
      required: [id, userId, title, slug, originalLanguage, status, supportedLanguages, createdAt, fields]
...
    FormTemplate:
      type: object
      properties:
        id:
          type: string
        formId:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        featureImageUrl:
          type: string
          nullable: true
        category:
          type: string
        isPublic:
          type: boolean
        useCount:
          type: integer
        createdAt:
          type: string
        fieldCount:
          type: integer
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FormField"
      required: [id, formId, userId, title, category, isPublic, useCount, createdAt, fieldCount]
    SaveTemplateBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        featureImageUrl:
          type: string
          nullable: true
        category:
          type: string
        isPublic:
          type: boolean
      required: [title, isPublic]